requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Pillow==10.0.0
python-dotenv==1.0.0
qrcode==7.4.2
WTForms==3.0.1
email-validator==2.0.0
gunicorn==21.2.0 